review: cut isolate_git_config duplication 4→2, trim test-code docs and lint-script comments
Addresses PR #755's three review comments.

1. "Why do we need multiple versions of this same function?"

   Investigated whether the src/tests compilation boundary is really the
   floor. It's not quite: spelunk-core already exposes test-only code
   across that exact boundary via `#[cfg(any(test, feature =
   "test-support"))]` (see `config::secret_store::MemoryStore`, already
   consumed today from spelunk-cli's own src-side `#[cfg(test)]` unit
   tests via a cross-crate dev-dependency). Applying that pattern here:
   `isolate_git_config`/`git_command` now live once in the new
   `spelunk-core/src/test_support.rs`, and spelunk-cli's two copies
   (`tests/plumbing_helpers.rs`, `src/cli/cmd/test_support.rs`) are gone,
   replaced with a direct re-export of spelunk-core's definition, the same
   proven cross-crate pattern MemoryStore already uses. That's 4 → 2.

   Tried to go further, to spelunk-core's own two remaining copies
   (src/test_support.rs + tests/common/mod.rs), via the same feature
   gate plus a self-referencing dev-dependency
   (`spelunk-core = { path = ".", features = ["test-support"] }`), the
   standard trick for letting a crate's own tests/ integration binaries
   see feature-gated src/ items. It compiles and passes in a clean
   target dir, but reproducibly breaks this repo's actual commit
   workflow: the pre-commit hook (`.git/hooks/pre-commit`) points
   `CARGO_TARGET_DIR` at the main checkout's shared target/ (to avoid
   multi-GB-per-worktree disk bloat), and building against that shared,
   cross-Cargo.lock target dir fails with `error[E0432]: unresolved
   import spelunk_core::test_support` — non-transient, reproduced twice.
   Reverted that last step as unsafe to ship; 2 total (one per src/tests
   side of spelunk-core) is the real, verified floor, not 2-per-crate (4).

2. Removed the `///` doc-comment block along with the
   tests/common/mod.rs copy of isolate_git_config it was attached to;
   its replacement copy (still needed per the finding above) uses plain
   `//` comments, per this repo's "no doc-comments in test code"
   convention.

3. Trimmed scripts/check-git-isolation.sh's comments from 94 lines to
   49 (267 to 223 total lines): kept the two scope rules, the "keep the
   test module last" invariant, the documented dataflow blind spot, and
   one line per self-test fixture naming the evasion it catches. Cut
   restated mechanism narration already documented at the
   isolate_git_config definitions themselves.

docs/testing.md's "four call sites" table is now stale (now three:
spelunk-core's src/test_support.rs + tests/common/mod.rs, plus
spelunk-cli's test_support.rs as a re-export, not a copy). Out of scope
for engineer per persona instructions; flagged for docs-writer.

Verification: cargo test -p spelunk-core -p spelunk-cli (exit 0, all
green), cargo fmt --all -- --check (clean), cargo clippy --all-targets
--features rich-formats -- -D warnings (clean, checked against both the
worktree's own target dir and the shared main-checkout target dir the
pre-commit hook uses), and scripts/check-git-isolation.sh --self-test /
check-git-isolation.sh . both OK on the trimmed script.
