+++ crates/spelunk-core/src/lib.rs
@@ -7,4 +7,6 @@ pub mod llm;
 pub mod registry;
 pub mod search;
 pub mod storage;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
 pub mod utils;
